dom/os/test/test_fopen.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>os fopen test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1141021">os tests</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>

  <script>
  var jsFuns = SpecialPowers.Cu.getJSTestingFunctions();

  function runTest(fn) {
    return new Promise((res, rej) => {
      var workerCode = `
        try {
          var file = navigator.os.fopen("../../../../dom/os/test/test_fopen.html", "rb");
          var cr = navigator.os.fclose(file);

          postMessage(JSON.stringify({
            hasFile: !!file,
            closeRet: cr
          }));
        }
        catch (ex) {
          postMessage("fopen err: " + ex);
        }
        `;

      var workerBlob = new Blob([workerCode], { type: 'application/javascript' });
      var w = new Worker(URL.createObjectURL(workerBlob));

      w.onmessage = e => {
        ok(e.data.indexOf('fopen err') === -1, 'ran successfully');

        try {
          var data = JSON.parse(e.data);
        }
        catch(ex) {
          ok(null, ex, fn + ' is valid JSON');
        }

        ok(data.hasFile === true, 'hasFile');
        ok(data.closeRet === 0, 'closeRet (exp. 0, was ' + data.closeRet + ')');

        res();
      };
    });
  }

  SimpleTest.waitForExplicitFinish();

  Promise.all([
    runTest()
  ]).then(() => {
    SimpleTest.finish();
  });
  </script>
</body>
</html>
